refining code, removing redundant modules
