--- src/service/controllers/jitter-controller.go
@@ -1,10 +1,0 @@
-package controllers
-
-import (
-	core "github.com/zairza-cetb/bench-routes/src/lib"
-)
-
-// JitterController controllers the jitter requests and transfers to the respective handler
-func JitterController(sig string) bool {
-	return core.HandlerJitterGeneral(sig)
-}
